+++ src/basic/__tests__/basic.test.tsx
@@ -5,7 +5,7 @@ import { CartPage } from '../../refactoring/pages/CartPage';
 import { AdminPage } from '../../refactoring/pages/AdminPage';
 import { CartItem, Coupon, Product } from '../../types';
 import { useCart, useCoupons, useProducts } from '../../refactoring/hooks';
-import * as cartUtils from '../../refactoring/hooks/utils/cartUtils';
+import * as cartUtils from '../../refactoring/utils/cartUtils';
 
 const mockProducts: Product[] = [
   {
