+++ src/refactoring/hooks/useCart.ts
@@ -1,7 +1,7 @@
 // useCart.ts
 import { useState } from 'react';
 import { CartItem, Coupon, Product } from '../../types';
-import { calculateCartTotal, getRemainingStock, updateCartItemQuantity } from './utils/cartUtils';
+import { calculateCartTotal, getRemainingStock, updateCartItemQuantity } from '../utils/cartUtils';
 
 export const useCart = () => {
   const [cart, setCart] = useState<CartItem[]>([]);
