+++ src/refactoring/pages/CartPage.tsx
@@ -1,6 +1,6 @@
 import { Coupon, Product } from '../../types.ts';
 import { useCart } from '../hooks/index.ts';
-import { getAppliedDiscount, getRemainingStock } from '../hooks/utils/cartUtils.ts';
+import { getAppliedDiscount, getRemainingStock } from '../utils/cartUtils.ts';
 import { getMaxDiscount } from '../utils/discountUtils.ts';
 
 interface Props {
