+++ src/refactoring/utils/cartUtils.ts
@@ -1,4 +1,4 @@
-import { CartItem, Coupon, Product } from '../../../types';
+import { CartItem, Coupon, Product } from '../../types';
 
 export const getMaxApplicableDiscount = (item: CartItem) => {
   const { product, quantity } = item;
